#namespace: Update membership create API to include manager relationships.

+++ pkg/infra/container_test.go
@@ -30,7 +30,7 @@ func TestContainer(t *testing.T) {
 	ass.NoError(err)
 	ass.NotNil(sv)
 	ass.Equal("128h0m0s", container.Modules.Access.SessionTimeout.String())
-	ass.Equal(uint16(100), container.Modules.Namespace.Manager.MaxNumberOfManager)
+	ass.Equal(100, container.Modules.Namespace.Manager.MaxNumberOfManager)
 }
 
 func Test_Request_JWT(t *testing.T) {
